[refactor] UserAction\Generic to Model\Action
